Fixed a mistake in the string-only chaincode
The value parameter needs to be declared as a string, not a []byte, even
though these are supposed to be equivalent in Go. The intermediate
layers in HLF treat all parameters as strings and can only convert them
to other types if there is a JSON serialization/deserialization defined
for them. That was the whole point of trying this setup instead of the
"blob" chaincode with []byte data.