Made the peer's hostname a command line option
The TLS certificate used to authenticate the gateway connection is
specific to the peer, not just the organization (like the other
certificates), so the client needs to use a certificate for peer1 or
peer2 if it's connecting to peer1 or peer2 instead of peer0. So both the
IP address and the name of the peer need to be set on the command line.