Fixed a bug in putString.js, fixed network-conf to use a path
The "length" property of strings is lowercase, not capitalized, but I
didn't notice because attempting to read a nonexistant Length property
just fails silently instead of causing a visible error. Also it turns
out it's possible for network-conf.yaml to refer to the CA certificate
by path instead of embedding the encoded certificate in the file (which
all the examples do).