Added an optional argument for message rate
For the "latency" test, we'll use the fixed-rate Caliper workload
module, but there's no way for the putString module to know what
workload module it's being called with. Adding a "rate" argument, and
manually configuring it to be the same rate as given to the fixed-rate
workload module, is a workaround for this.